Change pointers to shared pointers, Vectors of references to vectors of shared pointers

+++ simForest/simForest/GroundSquare.h
@@ -16,6 +16,6 @@ private:
 	BaseTree* mainTree;
 	BaseTree* growingTree;
 	Terrain terrain;
-	double acorns;
+	int acorns;
 
 };
